+++ seed/csharp-sdk/custom-auth/src/SeedCustomAuth.Test/Unit/MockServer/GetWithCustomAuthTest.cs
@@ -1,0 +1,36 @@
+using System.Threading.Tasks;
+using FluentAssertions.Json;
+using Newtonsoft.Json.Linq;
+using NUnit.Framework;
+using SeedCustomAuth.Core;
+
+#nullable enable
+
+namespace SeedCustomAuth.Test.Unit.MockServer;
+
+[TestFixture]
+public class GetWithCustomAuthTest : BaseMockServerTest
+{
+    [Test]
+    public async Task MockServerTest()
+    {
+        const string mockResponse = """
+            true
+            """;
+
+        Server
+            .Given(WireMock.RequestBuilders.Request.Create().WithPath("/custom-auth").UsingGet())
+            .RespondWith(
+                WireMock
+                    .ResponseBuilders.Response.Create()
+                    .WithStatusCode(200)
+                    .WithBody(mockResponse)
+            );
+
+        var response = await Client.CustomAuth.GetWithCustomAuthAsync(RequestOptions);
+        JToken
+            .Parse(mockResponse)
+            .Should()
+            .BeEquivalentTo(JToken.Parse(JsonUtils.Serialize(response)));
+    }
+}