+++ seed/csharp-sdk/examples/src/SeedExamples.Test/Unit/Serialization/ExtendedMovieTest.cs
@@ -18,28 +18,28 @@ public class ExtendedMovieTest
         var inputJson =
             @"
         {
-  ""id"": ""movie-sda231x"",
-  ""title"": ""Pulp Fiction"",
-  ""from"": ""Quentin Tarantino"",
-  ""rating"": 8.5,
-  ""type"": ""movie"",
-  ""tag"": ""tag-12efs9dv"",
-  ""cast"": [
-    ""John Travolta"",
-    ""Samuel L. Jackson"",
-    ""Uma Thurman"",
-    ""Bruce Willis""
-  ],
-  ""metadata"": {
-    ""academyAward"": true,
-    ""releaseDate"": ""2023-12-08"",
-    ""ratings"": {
-      ""rottenTomatoes"": 97,
-      ""imdb"": 7.6
-    }
-  }
-}
-";
+          ""id"": ""movie-sda231x"",
+          ""title"": ""Pulp Fiction"",
+          ""from"": ""Quentin Tarantino"",
+          ""rating"": 8.5,
+          ""type"": ""movie"",
+          ""tag"": ""tag-12efs9dv"",
+          ""cast"": [
+            ""John Travolta"",
+            ""Samuel L. Jackson"",
+            ""Uma Thurman"",
+            ""Bruce Willis""
+          ],
+          ""metadata"": {
+            ""academyAward"": true,
+            ""releaseDate"": ""2023-12-08"",
+            ""ratings"": {
+              ""rottenTomatoes"": 97,
+              ""imdb"": 7.6
+            }
+          }
+        }
+        ";
 
         var serializerOptions = new JsonSerializerOptions
         {