+++ seed/csharp-sdk/literal/src/SeedLiteral/Headers/HeadersClient.cs
@@ -27,12 +27,14 @@ public partial class HeadersClient
             { "X-Endpoint-Version", request.EndpointVersion.ToString() },
             { "X-Async", request.Async.ToString() },
         };
+        var requestBody = new Dictionary<string, object>() { { "query", request.Query } };
         var response = await _client.MakeRequestAsync(
             new RawClient.JsonApiRequest
             {
                 BaseUrl = _client.Options.BaseUrl,
                 Method = HttpMethod.Post,
                 Path = "headers",
+                Body = requestBody,
                 Headers = _headers,
                 Options = options,
             },