+++ seed/csharp-sdk/query-parameters/src/SeedQueryParameters/Core/Public/SeedQueryParametersApiException.cs
@@ -9,10 +9,10 @@ public class SeedQueryParametersApiException(string message, int statusCode, obj
     /// <summary>
     /// The error code of the response that triggered the exception.
     /// </summary>
-    public int StatusCode { get; } = statusCode;
+    public int StatusCode => statusCode;
 
     /// <summary>
     /// The body of the response that triggered the exception.
     /// </summary>
-    public object Body { get; } = body;
+    public object Body => body;
 }